requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.4
django-autoslug==1.9.8
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
iniconfig==1.1.1
isort==5.8.0
mccabe==0.6.1
packaging==20.9
pep8-naming==0.11.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
python-slugify==5.0.2
pytz==2021.1
sqlparse==0.4.1
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
